fix: errors

+++ src/main/java/lang/print/gaps/task5/NumberDivider.java
@@ -2,7 +2,7 @@ package lang.print.gaps.task5;
 
 public class NumberDivider {
     void divide(int toBeDivided) {
-        float result = toBeDivided / 5;
+        float result = (float)toBeDivided / 5;
         System.out.println(result);
     }
 }
