+++ src/main/java/lang/print/gaps/task5/TimeConvertor.java
@@ -2,7 +2,7 @@ package lang.print.gaps.task5;
 
 public class TimeConvertor {
     public void convert(float minutes) {
-        int seconds = (int) (minutes * 60);
+        float seconds = (int) (minutes * 60);
         System.out.println(seconds);
     }
 }
